Bugfix/no client plugin (#1)
* Fixed version number in the Torch plugin's manifest.xml

* Mod files required for publishing

* Code analysis fixes (minor)

* Fixed welding by client without the plugin installed

* Deployment batch files; Simplified linked folders

* Hiding errors due to attempting to re-create folders

* Storing run configurations in the repository

* Reordered run configurations

* 0.1.19

* Fixed build action on the Torch plugins's manifest.xml

+++ MultigridProjectorServer/Patches/MyProjectorBase_BuildInternal.cs
@@ -64,7 +64,7 @@ namespace MultigridProjectorServer
                     return true;
 
                 // We use the builtBy field to pass the subgrid index
-                projection.BuildInternal(cubeBlockPosition, owner, builder, requestInstant, (int) builtBy);
+                projection.BuildInternal(cubeBlockPosition, owner, builder, requestInstant, builtBy);
             }
             catch (Exception e)
             {
